Split  kanji search bar into component
